Use a non-recursive Makefile.am to build the sources.
The non-recursive Makefile.am is used to build the convenience library,
the main application and the panel plugin. This reduces the build time
and it is no longer necessary to clean the whole project before build
to test changes in the convenience library.

Recursive makefiles are still used for the translations, the
documentation and the icons.