Simplify (screenshooter_take_and_output_screenshot).
Now, the main dialog of the application is created once in main.c
and hidden/shown when necessary. The idle func is now only responsible
for taking the screenshot and hidding/showing the dialog when needed.

Add a dialog member to the ScreenshotData struct, set it correctly in
the panel plugin and in the application. Fix some indentation in the
panel plugin code.