Rework the actions dialog to add horodating.
Add two new options: the title of the capture is now customizable and
is used to generate the name of the saved captures. It is now also
possible to append the date and the time of capturing to the name of
the saved file (TODO: check if "horodating" exists in English).

Fix a leak in the panel plugin.