First implementation of a nice rubber banding.
When the screen is composited, use a nicer area selection mecanism. Draw
a fullscreen transparent window which is darker than the normal view.
When the user draws a rectangle with the mouse, show the "normal" view
in this rectangle: the area which will be grabbed is normal, the rest
of the screen is darker.

A "traditionnal" rubber banding ala Thunar or Xfdesktop could be nicer,
this has to be investigated.