Rework the region selection when not composited.
As spotted by Gauvain Pocentek, the "while" loop in the previous code
was not the best thing in the world as it lead to 100% CPU usage during
the selection. Instead, we use a filter function. This needs testing and
it would be great to share some code with the composited version.