Implement alternate fix for crash on upload (#11879)
The original fix complicated the code by individually allocating GValue
objects and then storing pointers to them in the job parameters array.

While the crash on upload was fixed, the program would still crash on
exit due to the clear_func used for the array (g_free); as per GLib's
documentation, "clear_func is expected to clear the contents of the
array element it is given, but not free the element itself".

The minimal fix in this commit should fix the original issue, and also
the crash on program exit (bug #13684).