Allow and remember 0sec delay for all regions (Bug #13763)
This changeset also permits the user to define a delay for the
"Select a region" mode via UI.
Note: if the first dialog is shown, i.e. a region was not passed
as an argument, a minimal delay of 200ms will be used to avoid
that dialog to appear on the screenshot. The same minimal delay
is also used for "Select a region" mode if no delay is defined.