Fix screenshot finalize behavior for plugin
screenshooter_take_screenshot used to be blocking but isn't for some
time already therefore we need a callback to asynchronously notify when
the screenshot action was completed or cancelled so the plugin button
can be unblocked, right now it was never blocking because that function
returned immediately.

This also simplifies the code by removing conditionals.